Ihre Commit-Nachricht

+++ src/s3_operations.py
@@ -173,7 +173,7 @@ def add_inventory_data(summary_data, all_data):
         
         # Fill NaN values with 0
         for col in ['InitialQuantity', 'SupplierDelivery_Delivered', 'SupplierDelivery_Planned']:
-            summary_data[col] = summary_data[col].fillna(0)
+            summary_data[col] = summary_data[col].fillna(0).astype(float)
         
         # Ensure Date column is datetime
         summary_data['Date'] = pd.to_datetime(summary_data['Date'])
